requirements.txt
fastapi
uvicorn
python-dotenv
openai
tiktoken
azure-storage-blob
python-multipart
azure-ai-documentintelligence
